+++ src/disk/partition.rs
@@ -1,8 +1,13 @@
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
 use std::path::{Path, PathBuf};
 
 use crate::disk::{partitions, FileSystem};
 use crate::Result;
 
+#[cfg_attr(feature = "serde", serde(crate = "renamed_serde"))]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Clone, Debug)]
 pub struct Partition {
 	pub(crate) device: String,