+++ src/network/net_if_stats.rs
@@ -1,11 +1,18 @@
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
 use crate::Bytes;
 
+#[cfg_attr(feature = "serde", serde(crate = "renamed_serde"))]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub enum Duplex {
 	Full,
 	Half,
 	Unknown,
 }
 
+#[cfg_attr(feature = "serde", serde(crate = "renamed_serde"))]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct NetIfStats {}
 
 impl NetIfStats {