+++ src/process/os/macos/kinfo.rs
@@ -6,7 +6,7 @@ use std::mem;
 use std::ptr;
 
 use mach::{boolean, vm_types};
-use nix::libc;
+use nix::{errno, libc};
 
 use crate::process::{io_error_to_process_error, ProcessError, ProcessResult};
 use crate::Pid;
@@ -140,13 +140,14 @@ pub struct vmspace {
 pub fn kinfo_processes() -> io::Result<Vec<kinfo_proc>> {
 	let mut name: [i32; 3] = [libc::CTL_KERN, libc::KERN_PROC, libc::KERN_PROC_ALL];
 	let mut size: libc::size_t = 0;
-	let mut processes: Vec<kinfo_proc> = vec![];
+	let mut processes: Vec<kinfo_proc> = Vec::new();
 
 	loop {
+		// Dry-run to get the size required for the process list
 		let result = unsafe {
 			libc::sysctl(
 				name.as_mut_ptr(),
-				3,
+				name.len() as libc::c_uint,
 				ptr::null_mut(),
 				&mut size,
 				ptr::null_mut(),
@@ -157,30 +158,45 @@ pub fn kinfo_processes() -> io::Result<Vec<kinfo_proc>> {
 			return Err(io::Error::last_os_error());
 		}
 
-		processes.reserve(size);
+		// Reserve enough room to store the whole process list.
+		// `size` is the number of bytes, not the number of processes.
+		let num_processes = size / mem::size_of::<kinfo_proc>();
+		if num_processes > processes.capacity() {
+			processes.reserve_exact(num_processes - processes.capacity());
+		}
 
+		// Attempt to store the process list in `processes`
 		let result = unsafe {
 			libc::sysctl(
 				name.as_mut_ptr(),
-				3,
+				name.len() as libc::c_uint,
 				processes.as_mut_ptr() as *mut libc::c_void,
 				&mut size,
 				ptr::null_mut(),
 				0,
 			)
 		};
-		match result {
-			libc::ENOMEM => continue,
-			code if code < 0 => return Err(io::Error::last_os_error()),
-			_ => {
-				let length = size / mem::size_of::<kinfo_proc>();
-				unsafe {
-					processes.set_len(length);
-				}
-				debug_assert!(!processes.is_empty());
 
-				return Ok(processes);
+		if result < 0 {
+			// Need to check to see if the error was due to `libc::ENOMEM`, this indicates that
+			// there was not enough space in `processes` to store the whole process list which can
+			// occur when a new process spawns between getting the size and storing. In this case
+			// we can simply try again.
+			if libc::ENOMEM == errno::errno() {
+				continue;
+			} else {
+				return Err(io::Error::last_os_error());
 			}
+		} else {
+			// Getting the list succeeded so let `processes` know how many processes it holds.
+			// Have to recompute `num_processes` since `size` may have changed.
+			let num_processes = size / mem::size_of::<kinfo_proc>();
+			unsafe {
+				processes.set_len(num_processes);
+			}
+			debug_assert!(!processes.is_empty());
+
+			return Ok(processes);
 		}
 	}
 }
@@ -198,7 +214,7 @@ pub fn kinfo_process(pid: Pid) -> ProcessResult<kinfo_proc> {
 	let result = unsafe {
 		libc::sysctl(
 			name.as_mut_ptr(),
-			4,
+			name.len() as libc::c_uint,
 			info.as_mut_ptr() as *mut libc::c_void,
 			&mut size,
 			ptr::null_mut(),