+++ CHANGELOG.md
@@ -15,6 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Make all public types serde Serializable and Deserializable
+
+### Fix
+
+- [process][macos] fix macos process kinfo "Cannot allocate memory" errors
+
 ## [v3.1.0] - 2020-05-10
 
 ### Added